Should've fixed problems
Version number now updated (I'll be sure to keep an eye out for that in
the future. I was unaware that was a thing).

The package is com.rowan.ieee.sac2014
That's what it was before, yes? I don't know why an issue has come up. I
confirmed it, and am updating the manifest and apk so if it doesn't work
that's definitely an issue.

Now entirely working off the RowanSAC build.